fix name collisions
